refactor(core): blogpost fragment colocation
